fixd bug getter separate getter file names
